job_matching_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
spacy==3.7.2
python-docx==1.0.1
pydantic==2.4.2
numpy==1.24.3
scikit-learn==1.3.2 